[IMP] Make canvas entire viewport, move instructions out of the way
Previously, the canvas didn't take up the whole viewport to leave space
for the instructions, this cause the pause "grey filter" to have a
visible demaraction at the top of the screen. This commit moves the
instructions to an absolute position in the bottom left of the screen,
which is less intrusive, and allows the canvas to take the entire
screen. It also removes the unused export on the Canvas constructor
function, as it's now only used internally. It also removes the width
and height parameters from the Canvas constructor function and replaces
the use of globals (width and height) with the window's inner size.