[IMP] fix eslint parser configuration, add eslint to dev dependencies
Previously, the parser was failing on async methods, I tried changing
the ecmaVersion in the configuration to 2017 to no avail, in the end I
switched the parser to esprima, which fixed the issue.

It turns out that I put the ecmaVersion inside of the ecmaFeatures
object inside of the parserOptions, when it should have been at the root
of the parserOptions, which was failing silently. It would seem that it
worked with esprima, because it was similarly ignored, but esprima was
parsing in ES8 mode by default.